+++ spec/health_spec.rb
@@ -197,6 +197,14 @@ describe Diplomat::Health do
       expect(health.service('foobar', options: options).first['Node']['Node']).to eq('foobar')
     end
 
+    it 'check service with multiple tag options' do
+      stub_request(:get, "http://localhost:8500/v1/health/service/foobar?tag=v1&tag=v2")
+        .and_return(body: json)
+      health = Diplomat::Health
+
+      expect(health.service('foobar', tag: ['v1', 'v2']).first['Node']['Node']).to eq('foobar')
+    end
+
     it 'should check service with node-meta option' do
       faraday.stub(:get).and_return(OpenStruct.new(body: json))
       health = Diplomat::Health.new(faraday)